requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0

# Retry and resilience
tenacity>=8.2.0

# LLM Provider SDKs
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0

# LangChain integrations (optional)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Monitoring and metrics (optional)
prometheus-client>=0.16.0

# Async support (optional)
aiohttp>=3.8.0
httpx>=0.24.0